requirements.txt
selenium==3.14.0
behave==1.2.6
allure-behave==2.5.1
requests==2.19.1
xmljson==0.1.9
lxml==4.2.5
PyYAML==3.13